pyproject.toml
[tool.poetry]
name = "chat-room"
version = "0.1.0"
description = ""
authors = ["mtr8650"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.1"
uvicorn = "^0.34.0"
motor = "^3.2.0"
beanie = "^1.22.0"
httpx = "^0.28.0"
trio = "^0.29.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^2.0"}
python-jose = "^3.4.0"
python-dotenv = "^1.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.23.0"
httpx = "^0.28.1"

